change component name  QandA to QACard as it's more sensible, make some changes in this component:transfer sorting and limiting displayed q-a information in it's parent component, so now QACard looks pure individual  entry
