+++ client/src/components/QuestionAnswerPart/LoadMoreAnswers.jsx
@@ -1,0 +1,10 @@
+import React, { useState } from 'react';
+//import PropTypes from 'prop-types';
+
+export default function LoadMoreAnswers() {
+  return (
+    <div>
+      <button type="button">LOAD MORE ANSWERS</button>
+    </div>
+  );
+}
