+++ client/src/components/QuestionsAndAnswers.jsx
@@ -4,12 +4,14 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import PropTypes from 'prop-types';
 import SearchBar from './QuestionAnswerPart/SearchBar.jsx';
-import QandA from './QuestionAnswerPart/QandA.jsx';
+import QACard from './QuestionAnswerPart/QACard.jsx';
+import MoreAandQ from './QuestionAnswerPart/MoreAandQ.jsx';
+import QuestionModal from './QuestionAnswerPart/QuestionModal.jsx';
 
 const configobj = require('../../../config.js');
 
 export default function QuestionsAndAnswers({ productId, productName }) {
-  const [qalist, setQalist] = useState({});
+  const [qalist, setQalist] = useState([]);
   const [pageNum, setPageNum] = useState(1);
 
   const url = 'https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfp/';
@@ -32,7 +34,7 @@ export default function QuestionsAndAnswers({ productId, productName }) {
   useEffect(() => {
     axios(options)
       .then((response) => {
-        setQalist(response.data);
+        setQalist(response.data.results);
         setPageNum(pageNum + 1);
       })
       .catch((err) => console.log('Error during get request on Q/A part for questions and answers list'));
@@ -42,8 +44,18 @@ export default function QuestionsAndAnswers({ productId, productName }) {
     <div>
       <div><h3 style={{ textAlign: 'left' }}>QUESTIONS & ANSWERS</h3></div>
       <SearchBar />
-      <QandA qalist={qalist} productName={productName} />
+      {qalist.length > 0 ? (qalist.map((ele, index) => (
+        <QACard
+          ele={ele}
+          key={index}
+          productName={productName}
+        />
+      )))
+        : <QuestionModal productId={qalist.product_id} productName={productName} />}
       <br />
+      <MoreAandQ />
+      <br />
+      <QuestionModal productId={qalist.product_id} productName={productName} />
     </div>
 
   );
